requirements.txt
APScheduler==3.6.3
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
dataclasses==0.8
decorator==5.0.9
idna==2.10
importlib-metadata==4.0.1
ipdb==0.13.7
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
requests==2.25.1
six==1.16.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==2.0.1
zipp==3.4.1
